feat(core): move file parsing to separate file
metriport/metriport-internal#1040
