+++ packages/core/src/external/surescripts/command/receive-response/receive-response-direct.ts
@@ -2,6 +2,8 @@ import { SurescriptsReceiveResponseHandler } from "./receive-response";
 import { SurescriptsSftpClient } from "../../client";
 import { SurescriptsFileIdentifier } from "../../types";
 
+// import { convertFlatFile } from "../../fhir-converter";
+
 export class SurescriptsReceiveResponseHandlerDirect implements SurescriptsReceiveResponseHandler {
   constructor(private readonly client: SurescriptsSftpClient = new SurescriptsSftpClient()) {}
 
@@ -11,9 +13,7 @@ export class SurescriptsReceiveResponseHandlerDirect implements SurescriptsRecei
   }: SurescriptsFileIdentifier): Promise<void> {
     const responseFile = await this.client.receiveResponse({ transmissionId, populationId });
     if (responseFile) {
-      console.log(
-        "TODO: ENG-377 - parse response file into FHIR bundle and place in conversion bucket"
-      );
+      responseFile.content;
     }
   }
 }
